refactor(order): Combine "Order Service" and add verification logic
- Combine "Order Service" and add verification logic
- Adjust data returns for DTO

+++ src/main/java/com/multicampus/gamesungcoding/a11ymarketserver/admin/order/model/AdminOrderRespDTO.java
@@ -21,7 +21,7 @@ public class AdminOrderRespDTO {
     private String receiverAddr1;
     private String receiverAddr2;
     private OrderStatus orderStatus;
-    private Long totalPrice;
+    private Integer totalPrice;
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
     private LocalDateTime createdAt;
 }
